fix(chat):随chat更新用户状态

+++ src/main/java/com/example/heart_field/dto/consultant/profile/ConsultantProfileDto.java
@@ -12,6 +12,6 @@ public class ConsultantProfileDto {
     private Integer maxCount;
     private Integer maxConcurrentCount;
     private Integer currentSessionCount;
-    private Integer todayTotalTime;
+    private Long todayTotalTime;
     private Integer todayTotalCount;
 }
